backend/requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.0
pydantic-settings==2.6.0

# Task Queue
celery==5.4.0
redis==5.2.0

# AI & LLM
langchain==0.2.16
langgraph==0.2.0
langchain-openai==0.1.25
langchain-groq==0.1.9
langchain-core==0.2.43
openai==1.54.0

# Code Analysis
tree-sitter==0.23.2
tree-sitter-python==0.23.6
tree-sitter-javascript==0.23.1
gitpython==3.1.41

# AWS
boto3==1.34.34

# Utilities
python-dotenv==1.0.0
httpx==0.26.0
aiofiles==23.2.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
